move tool to cmd-dev + better name
this is a tool for MT devs, it should not be part of the release
artifacts or public documentation.
