+++ .gitignore
@@ -2,6 +2,7 @@
 .notouch
 .vscode
 /build
+/cmd-dev/mt-simulate-memory-idx-lock-contention/mt-simulate-memory-idx-lock-contention
 /cmd/metrictank/metrictank
 /cmd/mt-aggs-explain/mt-aggs-explain
 /cmd/mt-explain/mt-explain
@@ -12,7 +13,6 @@
 /cmd/mt-kafka-mdm-sniff/mt-kafka-mdm-sniff
 /cmd/mt-kafka-persist-sniff/mt-kafka-persist-sniff
 /cmd/mt-schemas-explain/mt-schemas-explain
-/cmd/mt-simulate-lock-congestion/mt-simulate-lock-congestion
 /cmd/mt-split-metrics-by-ttl/mt-split-metrics-by-ttl
 /cmd/mt-store-cat/mt-store-cat
 /cmd/mt-store-cp-experimental/mt-store-cp-experimental
