+++ cmd-dev/mt-simulate-memory-idx-lock-contention/runner/stat.go
